[Feat]카드리스트 자료형, repository수정

+++ src/main/java/com/shinhan/soloplay/card/CardDTO.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class CardDTO {
-	private int cardId;
+	private Long cardId;
 	private String cardName;
 	private String cardType;
 	private String cardBenefit;
